ImageRecognitionAPI/requirements.txt
Flask==2.3.2
tensorflow==2.12.0
Pillow==10.0.0
sendgrid==6.11.0
Flask-Limiter==2.8.1